+++ src/Sulu/Bundle/ContactBundle/Admin/ContactContentNavigationProvider.php
@@ -13,7 +13,7 @@ namespace Sulu\Bundle\ContactBundle\Admin;
 use Sulu\Bundle\AdminBundle\Navigation\ContentNavigationItem;
 use Sulu\Bundle\AdminBundle\Navigation\ContentNavigationProviderInterface;
 
-class SuluContactContentNavigationProvider implements ContentNavigationProviderInterface
+class ContactContentNavigationProvider implements ContentNavigationProviderInterface
 {
     public function getNavigationItems(array $options = array())
     {