+++ src/Sulu/Bundle/WebsiteBundle/Admin/WebsiteAdmin.php
@@ -8,30 +8,25 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\TranslateBundle\Admin;
+namespace Sulu\Bundle\WebsiteBundle\Admin;
 
 use Sulu\Bundle\AdminBundle\Admin\Admin;
 use Sulu\Bundle\AdminBundle\Navigation\Navigation;
 use Sulu\Bundle\AdminBundle\Navigation\NavigationItem;
-use Sulu\Bundle\TranslateBundle\Command\ExportCommand;
-use Sulu\Bundle\TranslateBundle\Command\ImportCommand;
 
-class SuluTranslateAdmin extends Admin
+class WebsiteAdmin extends Admin
 {
-
     public function __construct($title)
     {
         $rootNavigationItem = new NavigationItem($title);
-
         $section = new NavigationItem('');
 
         $settings = new NavigationItem('navigation.settings');
-        $settings->setIcon('cogwheels');
+        $settings->setIcon('gear');
 
-        $translate = new NavigationItem('navigation.settings.translate');
-        $translate->setAction('settings/translate');
-        $translate->setIcon('book-open');
-        $settings->addChild($translate);
+        $roles = new NavigationItem('navigation.settings.cache', $settings);
+        $roles->setAction('settings/cache');
+        $roles->setIcon('hdd-o');
 
         $section->addChild($settings);
         $rootNavigationItem->addChild($section);
@@ -43,10 +38,7 @@ class SuluTranslateAdmin extends Admin
      */
     public function getCommands()
     {
-        return array(
-            new ImportCommand(),
-            new ExportCommand()
-        );
+        return array();
     }
 
     /**
@@ -54,6 +46,6 @@ class SuluTranslateAdmin extends Admin
      */
     public function getJsBundleName()
     {
-        return 'sulutranslate';
+        return 'suluwebsite';
     }
 }