+++ CHANGELOG.md
@@ -2,7 +2,8 @@ CHANGELOG for Sulu
 ==================
 
 * dev-develop
-    * BUGFIX      #945 [WebsiteBundle]   Fix Redirect url with query string correctly and trailing slash
+    * BUGFIX      #945  [WebsiteBundle]  Fix Redirect url with query string correctly and trailing slash
+    * ENHANCEMENT #1029 [All]            Removed prefixes from content navigation providers and admins
 
 * 0.17.0 (2015-04-20)
     * BUGFIX      #1020 [ContactBundle]  Fixed organization go back type bug